Add Z package
This is heavily derived from fuzz.sty, but it contains a few
changes that make this package easier to use:
* Remove dependency on the METAFONT symbols from fuzz. This makes
  it a lot easier to distribute the style.
* Add \label for each schema
* Add \real symbol

It should maintain backwards compatibility with fuzz (i.e. a valid
fuzz document should compile fine with this), but since it contains
some extensions (e.g. \real, but more in the future) the reverse is
not true.